requirements.txt
pystray==0.19.3
keyboard==0.13.5
Pillow==9.0.1